refactoring into a single module
