+++ README.md
@@ -1,2 +1,4 @@
 # analysis-code
-generic analysis code development
+Development of analysis functions to analyse physical Problems.
+
+So far scripts are developed for the analysis of pipi, Kpi, KK and rho on the lattice.
