+++ analysis/corr_matrix.py
@@ -25,6 +25,8 @@
 #
 ################################################################################
 
+__all__ = ["create_corr_matrix", "calculate_gevp"]
+
 import os
 import numpy as np
 import scipy.linalg as spla
