+++ analysis/zeta/__init__.py
@@ -1,0 +1,5 @@
+from zeta import *
+
+__all__ = [t for t in dir() if not t.startswith('_')]
+
+del t
