+++ rho_analyse
@@ -30,11 +30,7 @@
 import os
 import numpy as np
 
-import input_output as io
-import bootstrap
-import corr_matrix as cm
-import fit
-import analyze_fcts as af
+import analysis as ana
 
 def print_results(data, error):
     """Prints the results to screen
@@ -65,15 +61,15 @@ def read_new_data(nbsamples, path, filelist, tmin, lattice, d2, verbose=False):
     # read in data for GEVP
     if verbose:
         print("reading data")
-    corr_mat = cm.create_corr_matrix(nbsamples, path, filelist)
-    io.write_data(corr_mat, "./raw_data/corr_mat_rho_%s_TP%d.npy"%(lattice,d2))
+    corr_mat = ana.create_corr_matrix(nbsamples, path, filelist)
+    ana.write_data(corr_mat, "./raw_data/corr_mat_rho_%s_TP%d.npy"%(lattice,d2))
     # solve the GEVP
     if verbose:
         print("solving GEVP")
-    gevp_mat = cm.calculate_gevp(corr_mat, tmin)
-    io.write_data(gevp_mat, "./raw_data/gevp_mat_rho_%s_TP%d.npy"%(lattice,d2))
+    gevp_mat = ana.calculate_gevp(corr_mat, tmin)
+    ana.write_data(gevp_mat, "./raw_data/gevp_mat_rho_%s_TP%d.npy"%(lattice,d2))
     # write in ASCII
-    io.write_data_ascii(gevp_mat,"./raw_data/gevp_rho_%s_TP%d.dat"%(lattice,d2))
+    ana.write_data_ascii(gevp_mat,"./raw_data/gevp_rho_%s_TP%d.dat"%(lattice,d2))
     return gevp_mat
 
 def fit_corr(data, lolist, uplist, fitfunc, start_params, tmin, lattice, d,
@@ -84,11 +80,11 @@ def fit_corr(data, lolist, uplist, fitfunc, start_params, tmin, lattice, d,
     path="./plots/"
     plotlabel="corr"
     d2 = np.dot(d, d)
-    res, chi2, pval = fit.genfit(data, lolist, uplist, fitfunc, start_params,
+    res, chi2, pval = ana.genfit(data, lolist, uplist, fitfunc, start_params,
         tmin, lattice, d, label, path, plotlabel, verbose)
-    io.write_data(res, "./raw_data/corr_fit_res_rho_%s_TP%d.npy"%(lattice,d2))
-    io.write_data(chi2, "./raw_data/corr_fit_chi2_rho_%s_TP%d.npy"%(lattice,d2))
-    io.write_data(pval, "./raw_data/corr_fit_pval_rho_%s_TP%d.npy"%(lattice,d2))
+    ana.write_data(res, "./raw_data/corr_fit_res_rho_%s_TP%d.npy"%(lattice,d2))
+    ana.write_data(chi2, "./raw_data/corr_fit_chi2_rho_%s_TP%d.npy"%(lattice,d2))
+    ana.write_data(pval, "./raw_data/corr_fit_pval_rho_%s_TP%d.npy"%(lattice,d2))
     return res, chi2, pval
 
 
@@ -100,12 +96,12 @@ def fit_mass(data, lolist, uplist, fitfunc, start_params, tmin, lattice, d,
     path="./plots/"
     plotlabel="mass"
     d2 = np.dot(d, d)
-    mass, mmass, dmass = af.compute_mass(data, False)
-    res, chi2, pval = fit.genfit(mass, lolist, uplist, fitfunc, start_params,
+    mass, mmass, dmass = ana.compute_mass(data, False)
+    res, chi2, pval = ana.genfit(mass, lolist, uplist, fitfunc, start_params,
         tmin, lattice, d, label, path, plotlabel, verbose)
-    io.write_data(res, "./raw_data/mass_fit_res_rho_%s_TP%d.npy"%(lattice,d2))
-    io.write_data(chi2, "./raw_data/mass_fit_chi2_rho_%s_TP%d.npy"%(lattice,d2))
-    io.write_data(pval, "./raw_data/mass_fit_pval_rho_%s_TP%d.npy"%(lattice,d2))
+    ana.write_data(res, "./raw_data/mass_fit_res_rho_%s_TP%d.npy"%(lattice,d2))
+    ana.write_data(chi2, "./raw_data/mass_fit_chi2_rho_%s_TP%d.npy"%(lattice,d2))
+    ana.write_data(pval, "./raw_data/mass_fit_pval_rho_%s_TP%d.npy"%(lattice,d2))
     return res, chi2, pval
 
 def read_fit_corr(lattice, d):
@@ -113,9 +109,9 @@ def read_fit_corr(lattice, d):
     """
     print("reading correlation fit data")
     d2 = np.dot(d, d)
-    res = io.read_data("./raw_data/corr_fit_res_rho_%s_TP%d.npy"%(lattice,d2))
-    chi2 = io.read_data("./raw_data/corr_fit_chi2_rho_%s_TP%d.npy"%(lattice,d2))
-    pval = io.read_data("./raw_data/corr_fit_pval_rho_%s_TP%d.npy"%(lattice,d2))
+    res = ana.read_data("./raw_data/corr_fit_res_rho_%s_TP%d.npy"%(lattice,d2))
+    chi2 = ana.read_data("./raw_data/corr_fit_chi2_rho_%s_TP%d.npy"%(lattice,d2))
+    pval = ana.read_data("./raw_data/corr_fit_pval_rho_%s_TP%d.npy"%(lattice,d2))
     return res, chi2, pval
 
 def read_fit_mass(lattice, d):
@@ -123,9 +119,9 @@ def read_fit_mass(lattice, d):
     """
     print("reading mass fit data")
     d2 = np.dot(d, d)
-    res = io.read_data("./raw_data/mass_fit_res_rho_%s_TP%d.npy"%(lattice,d2))
-    chi2 = io.read_data("./raw_data/mass_fit_chi2_rho_%s_TP%d.npy"%(lattice,d2))
-    pval = io.read_data("./raw_data/mass_fit_pval_rho_%s_TP%d.npy"%(lattice,d2))
+    res = ana.read_data("./raw_data/mass_fit_res_rho_%s_TP%d.npy"%(lattice,d2))
+    chi2 = ana.read_data("./raw_data/mass_fit_chi2_rho_%s_TP%d.npy"%(lattice,d2))
+    pval = ana.read_data("./raw_data/mass_fit_pval_rho_%s_TP%d.npy"%(lattice,d2))
     return res, chi2, pval
 
 def calc_phaseshift(data, L, d, lattice, mpi, verbose=True):
@@ -147,17 +143,17 @@ def calc_phaseshift(data, L, d, lattice, mpi, verbose=True):
 
     if verbose:
         print("calculating CM energy")
-    gamma, Ecm = af.calculate_cm_energy(data, L, d)
+    gamma, Ecm = ana.calc_Ecm(data, L, d)
     if verbose:
         print("calculating q^2")
-    q2 = af.calculate_q(Ecm, mpi, L)
+    q2 = ana.calculate_q(Ecm, mpi, L)
     if verbose:
         print("calculating delta")
-    delta, tandelta = af.calculate_delta(q2, gamma, d)
+    delta, tandelta = ana.calculate_delta(q2, gamma, d)
 
-    mdata[:,0], ddata[:,0] = af.return_mean_corr(Ecm)
-    mdata[:,1], ddata[:,1] = af.return_mean_corr(delta)
-    mdata[:,2], ddata[:,2] = af.return_mean_corr(tandelta)
+    mdata[:,0], ddata[:,0] = ana.return_mean_corr(Ecm)
+    mdata[:,1], ddata[:,1] = ana.return_mean_corr(delta)
+    mdata[:,2], ddata[:,2] = ana.return_mean_corr(tandelta)
     mdata[:,3] = np.sin(mdata[:,1])**2
     ddata[:,3] = np.abs(np.sin(mdata[:,1]*2.)*ddata[:,1])
     odata[:,0] = Ecm[0]
@@ -332,7 +328,7 @@ def analyse(lattice="A40.24", d=np.array([0., 0., 0.]), verbose=True):
     #gevp_mat = read_new_data(nbsamples, path, filelist, tmin, lattice, d2)
 
     ## read in solution of GEVP
-    gevp_mat = io.read_data("./raw_data/gevp_mat_rho_%s_TP%d.npy" % (lattice, d2))
+    gevp_mat = ana.read_data("./raw_data/gevp_mat_rho_%s_TP%d.npy" % (lattice, d2))
 
     ## fit correlation function
     res, chi2, pvals = fit_corr(gevp_mat, lolist, uplist, fitfunc, start_params,
@@ -347,7 +343,7 @@ def analyse(lattice="A40.24", d=np.array([0., 0., 0.]), verbose=True):
     #print(np.array_equal(massres, massres1))
 
     ## calculate phaseshift
-    #odata, mdata, ddata = calc_phaseshift(res[:,1], L, d, lattice, mpi)
+    odata, mdata, ddata = calc_phaseshift(res[:,1], L, d, lattice, mpi)
     return
 
 def main():
